# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SwitchPoint API",
    "description": "Implements the FCA-prescribed pension switching, DB transfer and cashflow analyses for UK IFAs. Authenticate with POST /api/v1/auth/login and send 'Authorization: Bearer <token>'.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5093/"
    }
  ],
  "paths": {
...
    "/api/v1/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfClientSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClientWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}/schemes": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}/schemes/{schemeId}": {
      "put": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "schemeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SchemeWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemeDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "schemeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}/analyses": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clients/{id}/reports": {
      "get": {
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddressDto": {
        "required": [
          "line1",
          "line2",
          "town",
          "county",
          "postcode",
          "country"
        ],
        "type": "object",
        "properties": {
          "line1": {
            "type": [
              "null",
              "string"
            ]
          },
          "line2": {
            "type": [
              "null",
              "string"
            ]
          },
          "town": {
            "type": [
              "null",
              "string"
            ]
          },
          "county": {
            "type": [
              "null",
              "string"
            ]
          },
          "postcode": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "AnalysisKind": {
        "type": "integer"
      },
...
      "AnalysisSummary": {
        "required": [
          "id",
          "kind",
          "title",
          "status",
          "version",
          "calculatedAtUtc",
          "updatedAtUtc",
          "createdBy",
          "clientId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/AnalysisKind"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "calculatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ClientDetail": {
        "required": [
          "id",
          "fullName",
          "age",
          "externalReference",
          "firstName",
          "lastName",
          "dateOfBirth"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nationalInsuranceNumberMasked": {
            "type": [
              "null",
              "string"
            ]
          },
          "externalReference": {
            "$ref": "#/components/schemas/ExternalReferenceDto"
          },
          "schemes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchemeDto"
            }
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "address": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "maritalStatus": {
            "$ref": "#/components/schemas/MaritalStatus"
          },
          "employmentStatus": {
            "$ref": "#/components/schemas/EmploymentStatus"
          },
          "annualSalary": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "targetRetirementAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "taxRegime": {
            "$ref": "#/components/schemas/TaxRegime"
          },
          "riskProfile": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "health": {
            "$ref": "#/components/schemas/HealthStatus"
          },
          "isSmoker": {
            "type": "boolean"
          },
          "statePension": {
            "$ref": "#/components/schemas/StatePensionDto"
          },
          "nationalInsuranceNumber": {
            "type": [
              "null",
              "string"
            ],
            "description": "Write-only; stored masked."
          }
        }
      },
      "ClientSummary": {
        "required": [
          "id",
          "fullName",
          "dateOfBirth",
          "age",
          "email",
          "riskProfile",
          "schemeCount",
          "totalPensionValue",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "riskProfile": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "schemeCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPensionValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClientWrite": {
        "required": [
          "firstName",
          "lastName",
          "dateOfBirth"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "address": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "maritalStatus": {
            "$ref": "#/components/schemas/MaritalStatus"
          },
          "employmentStatus": {
            "$ref": "#/components/schemas/EmploymentStatus"
          },
          "annualSalary": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "targetRetirementAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "taxRegime": {
            "$ref": "#/components/schemas/TaxRegime"
          },
          "riskProfile": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "health": {
            "$ref": "#/components/schemas/HealthStatus"
          },
          "isSmoker": {
            "type": "boolean"
          },
          "statePension": {
            "$ref": "#/components/schemas/StatePensionDto"
          },
          "nationalInsuranceNumber": {
            "type": [
              "null",
              "string"
            ],
            "description": "Write-only; stored masked."
          }
        }
      },
...
      "EmploymentStatus": {
        "type": "integer"
      },
...
      "ExternalReferenceDto": {
        "required": [
          "source",
          "externalId"
        ],
        "type": "object",
        "properties": {
          "source": {
            "$ref": "#/components/schemas/ExternalSource"
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "HealthStatus": {
        "type": "integer",
        "description": "Health status affecting annuity terms."
      },
...
      "MaritalStatus": {
        "type": "integer"
      },
...
      "PagedResultOfClientSummary": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReportDto": {
        "required": [
          "id",
          "clientId",
          "analysisId",
          "analysisVersion",
          "analysisResultHash",
          "kind",
          "format",
          "templateVersion",
          "generatedBy",
          "sha256",
          "sizeBytes",
          "generatedAtUtc",
          "downloadUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "analysisId": {
            "type": "string",
            "format": "uuid"
          },
          "analysisVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "analysisResultHash": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ReportKind"
          },
          "format": {
            "$ref": "#/components/schemas/ReportFormat"
          },
          "templateVersion": {
            "type": "string"
          },
          "generatedBy": {
            "type": "string",
            "format": "uuid"
          },
          "sha256": {
            "type": "string"
          },
          "sizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "generatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "downloadUrl": {
            "type": "string"
          }
        }
      },
...
      "SchemeDto": {
        "required": [
          "id",
          "clientId",
          "type",
          "productName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "providerName": {
            "type": [
              "null",
              "string"
            ]
          },
          "netTransferValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "weightedOcfPct": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/SchemeType"
          },
          "providerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "policyNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "currentValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "transferValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "valuationDate": {
            "type": "string",
            "format": "date"
          },
          "startDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "charges": {
            "$ref": "#/components/schemas/ChargeScheduleDto"
          },
          "guarantees": {
            "$ref": "#/components/schemas/GuaranteesDto"
          },
          "selectedRetirementAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "inDrawdown": {
            "type": "boolean"
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContributionDto"
            }
          },
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HoldingDto"
            }
          },
          "definedBenefit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DefinedBenefitDto"
              }
            ]
          }
        }
      },
...
      "SchemeType": {
        "type": "integer",
        "description": "The kind of existing arrangement a client holds."
      },
      "SchemeWrite": {
        "required": [
          "type",
          "productName"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/SchemeType"
          },
          "providerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "productName": {
            "type": "string"
          },
          "policyNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "currentValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "transferValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "valuationDate": {
            "type": "string",
            "format": "date"
          },
          "startDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "charges": {
            "$ref": "#/components/schemas/ChargeScheduleDto"
          },
          "guarantees": {
            "$ref": "#/components/schemas/GuaranteesDto"
          },
          "selectedRetirementAge": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "inDrawdown": {
            "type": "boolean"
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContributionDto"
            }
          },
          "holdings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HoldingDto"
            }
          },
          "definedBenefit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DefinedBenefitDto"
              }
            ]
          }
        }
      },
      "Sex": {
        "type": "integer"
      },
      "StatePensionDto": {
        "required": [
          "forecastWeeklyAmount",
          "qualifyingYears"
        ],
        "type": "object",
        "properties": {
          "forecastWeeklyAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "qualifyingYears": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "TaxRegime": {
        "type": "integer",
        "description": "Which income tax regime applies to the client's non-savings income."
      },
...
    }
  },
...
}